+++ staging/src/k8s.io/endpointslice/utils_test.go
@@ -1316,3 +1316,232 @@ func Test_hintsEnabled(t *testing.T) {
 		})
 	}
 }
+
+func TestFindPort(t *testing.T) {
+	restartAlways := v1.ContainerRestartPolicyAlways
+	testCases := []struct {
+		name           string
+		containers     []v1.Container
+		initContainers []v1.Container
+		port           intstr.IntOrString
+		expected       int
+		pass           bool
+	}{{
+		name:       "valid int, no ports",
+		containers: []v1.Container{{}},
+		port:       intstr.FromInt32(93),
+		expected:   93,
+		pass:       true,
+	}, {
+		name: "valid int, with ports",
+		containers: []v1.Container{{Ports: []v1.ContainerPort{{
+			Name:          "",
+			ContainerPort: 11,
+			Protocol:      "TCP",
+		}, {
+			Name:          "p",
+			ContainerPort: 22,
+			Protocol:      "TCP",
+		}}}},
+		port:     intstr.FromInt32(93),
+		expected: 93,
+		pass:     true,
+	}, {
+		name:       "valid str, no ports",
+		containers: []v1.Container{{}},
+		port:       intstr.FromString("p"),
+		expected:   0,
+		pass:       false,
+	}, {
+		name: "valid str, one ctr with ports",
+		containers: []v1.Container{{Ports: []v1.ContainerPort{{
+			Name:          "",
+			ContainerPort: 11,
+			Protocol:      "UDP",
+		}, {
+			Name:          "p",
+			ContainerPort: 22,
+			Protocol:      "TCP",
+		}, {
+			Name:          "q",
+			ContainerPort: 33,
+			Protocol:      "TCP",
+		}}}},
+		port:     intstr.FromString("q"),
+		expected: 33,
+		pass:     true,
+	}, {
+		name: "valid str, two ctr with ports",
+		containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+			Name:          "",
+			ContainerPort: 11,
+			Protocol:      "UDP",
+		}, {
+			Name:          "p",
+			ContainerPort: 22,
+			Protocol:      "TCP",
+		}, {
+			Name:          "q",
+			ContainerPort: 33,
+			Protocol:      "TCP",
+		}}}},
+		port:     intstr.FromString("q"),
+		expected: 33,
+		pass:     true,
+	}, {
+		name: "valid str, two ctr with same port",
+		containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+			Name:          "",
+			ContainerPort: 11,
+			Protocol:      "UDP",
+		}, {
+			Name:          "p",
+			ContainerPort: 22,
+			Protocol:      "TCP",
+		}, {
+			Name:          "q",
+			ContainerPort: 22,
+			Protocol:      "TCP",
+		}}}},
+		port:     intstr.FromString("q"),
+		expected: 22,
+		pass:     true,
+	}, {
+		name: "valid str, invalid protocol",
+		containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+			Name:          "a",
+			ContainerPort: 11,
+			Protocol:      "snmp",
+		},
+		}}},
+		port:     intstr.FromString("a"),
+		expected: 0,
+		pass:     false,
+	}, {
+		name: "valid hostPort",
+		containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+			Name:          "a",
+			ContainerPort: 11,
+			HostPort:      81,
+			Protocol:      "TCP",
+		},
+		}}},
+		port:     intstr.FromString("a"),
+		expected: 11,
+		pass:     true,
+	},
+		{
+			name: "invalid hostPort",
+			containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+				Name:          "a",
+				ContainerPort: 11,
+				HostPort:      -1,
+				Protocol:      "TCP",
+			},
+			}}},
+			port:     intstr.FromString("a"),
+			expected: 11,
+			pass:     true,
+			// this should fail but passes.
+		},
+		{
+			name: "invalid ContainerPort",
+			containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+				Name:          "a",
+				ContainerPort: -1,
+				Protocol:      "TCP",
+			},
+			}}},
+			port:     intstr.FromString("a"),
+			expected: -1,
+			pass:     true,
+			// this should fail but passes
+		},
+		{
+			name: "HostIP Address",
+			containers: []v1.Container{{}, {Ports: []v1.ContainerPort{{
+				Name:          "a",
+				ContainerPort: 11,
+				HostIP:        "192.168.1.1",
+				Protocol:      "TCP",
+			},
+			}}},
+			port:     intstr.FromString("a"),
+			expected: 11,
+			pass:     true,
+		},
+		{
+			name: "Sidecar initContainer named port",
+			initContainers: []v1.Container{{
+				RestartPolicy: &restartAlways,
+				Ports: []v1.ContainerPort{{
+					Name:          "a",
+					ContainerPort: 80,
+					HostPort:      -1,
+					Protocol:      "TCP",
+				}},
+			}},
+			port:     intstr.FromString("a"),
+			expected: 80,
+			pass:     true,
+		},
+		{
+			name: "Invalid(restartPolicy != Always) initContainer named port",
+			initContainers: []v1.Container{{
+				Ports: []v1.ContainerPort{{
+					Name:          "a",
+					ContainerPort: 80,
+					HostPort:      -1,
+					Protocol:      "TCP",
+				}},
+			}},
+			port:     intstr.FromString("a"),
+			expected: 0,
+			pass:     false,
+		},
+		{
+			name: "App and sidecar containers have the same named port, first app container port will be used",
+			initContainers: []v1.Container{{
+				RestartPolicy: &restartAlways,
+				Ports: []v1.ContainerPort{{
+					Name:          "a",
+					ContainerPort: 80,
+					HostPort:      -1,
+					Protocol:      "TCP",
+				}},
+			}},
+			containers: []v1.Container{{
+				Ports: []v1.ContainerPort{{
+					Name:          "a",
+					ContainerPort: 81,
+					HostPort:      -1,
+					Protocol:      "TCP",
+				}},
+			}, {
+				Ports: []v1.ContainerPort{{
+					Name:          "a",
+					ContainerPort: 82,
+					HostPort:      -1,
+					Protocol:      "TCP",
+				}},
+			}},
+			port:     intstr.FromString("a"),
+			expected: 81,
+			pass:     true,
+		},
+	}
+
+	for _, tc := range testCases {
+		port, err := FindPort(&v1.Pod{Spec: v1.PodSpec{Containers: tc.containers, InitContainers: tc.initContainers}},
+			&v1.ServicePort{Protocol: "TCP", TargetPort: tc.port})
+		if err != nil && tc.pass {
+			t.Errorf("unexpected error for %s: %v", tc.name, err)
+		}
+		if err == nil && !tc.pass {
+			t.Errorf("unexpected non-error for %s: %d", tc.name, port)
+		}
+		if port != tc.expected {
+			t.Errorf("wrong result for %s: expected %d, got %d", tc.name, tc.expected, port)
+		}
+	}
+}
